- bunch of minor tweaks from MSVC  - version bump    - rebuilt the configure script  - new libusb-win32 (but need a new package still)

+++ software/usb_ir/drivers/libusbpre1.c
@@ -272,7 +272,7 @@ static bool updateDeviceList(deviceList *devList)
 
                         /* basic stuff */
                         newDev->info.type = list->ids[pos];
-                        newDev->busIndex = busIndex;
+                        newDev->busIndex = (uint8_t)busIndex;
                         newDev->devIndex = LIBUSB_DEVNUM(dev);
 
                         /* determine the id (reusing if possible) */
